requirements.txt
djangorestframework=3.8.2
django=1.11.13
